added user logout feature
