+++ src/scripts/welcomePage/welcomeHtml.js
@@ -4,9 +4,6 @@ export default {
     <nav id="navBar">
         <ul class="topnav">
         <img class="left" src="src/images/otter1.png">
-        <li><a class="active" href="#home">Home</a></li>
-        <li><a href="#news">News</a></li>
-        <li><a href="#contact">Contact</a></li>
         </ul>
     </nav>
     <div id="welcome-page">
